kloud/resize: move resize functions to resize.go
